+++ fili-core/src/main/java/com/yahoo/bard/webservice/data/config/names/DataSourceName.java
@@ -7,6 +7,7 @@ import java.util.Comparator;
 /**
  * Marker interface for objects that can be treated as a data source name in druid.
  */
+@FunctionalInterface
 public interface DataSourceName {
 
     /**